Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@4126 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ play/src/Player.java
@@ -188,7 +188,7 @@ public class Player implements Immutable
     	for (int i = 0; i < rb.length; i++)
         {
             byte b = rb[i];
-            System.out.print(Integer.toHexString(b));
+            System.out.print(Byte.toHexString(b));
             System.out.print(' ');
         }
         System.out.println();
